Method for C0 parameter computation clarified/amplified.
Following text incorporated in the computation method section for C0:
"If a concentration in the profile at TIME<=0 appears AND its value
is numerically zero (0), that concentration data point is removed
from the generation of C0, the non-zero concentration value as the
value for the C0 parameter. This is consistent with scenario 1) below.

If no concentration-time data points appear in the profile at time of
last dose, then scenario 2) is utilized for estimating the value of C0."